+++ src/main/java/com/comp2024b/tocountornot/controller/CardController.java
@@ -2,6 +2,7 @@ package com.comp2024b.tocountornot.controller;
 
 import com.comp2024b.tocountornot.annotation.TokenRequired;
 import com.comp2024b.tocountornot.bean.Card;
+import com.comp2024b.tocountornot.service.UserService;
 import com.comp2024b.tocountornot.util.Result;
 import com.comp2024b.tocountornot.util.Results;
 import com.comp2024b.tocountornot.service.CardService;
@@ -12,36 +13,60 @@ import org.springframework.web.bind.annotation.*;
 public class CardController {
 
     private final CardService cardService;
+    private final UserService userService;
 
-    public CardController(CardService cardService) {
+    public CardController(CardService cardService, UserService userService) {
         this.cardService = cardService;
-    }
-
-    @TokenRequired
-    @GetMapping("{id}")
-    public Result selectCardById(@PathVariable("id") int id) {
-        Card card = cardService.selectCardById(id);
-        return Results.getSuccessResult(card);
-    }
-
-    @TokenRequired
-    @DeleteMapping("delete/{id}")
-    public Result deleteCard(@PathVariable("id") int id) {
-        cardService.deleteCard(id);
-        return Results.getSuccessResult(id);
+        this.userService = userService;
     }
 
     @TokenRequired
     @PostMapping("insert")
-    public Result insertCard(@RequestBody Card card) {
+    public Result insertCard(@RequestHeader("token") String token, @RequestBody Card card) {
+        card.setUid(userService.getUserIdWithToken(token));
         cardService.insertCard(card);
         return Results.getSuccessResult(card.getId());
     }
 
     @TokenRequired
-    @PutMapping("update")
-    public Result updateCard(@RequestBody Card card) {
-        cardService.updateCard(card);
-        return Results.getSuccessResult(card.getId());
+    @DeleteMapping("delete/{id}")
+    public Result deleteCard(@PathVariable("id") int id) {
+        Card c = cardService.selectCardById(id);
+        if (c == null) {
+            return Results.getNotFoundResult("Card does not exist");
+        } else {
+            cardService.deleteCard(id);
+            return Results.getSuccessResult();
+        }
+    }
+
+    @TokenRequired
+    @PutMapping("update/{id}")
+    public Result updateCard(@RequestHeader("token") String token, @RequestBody Card card, @PathVariable("id") int id) {
+        Card c = cardService.selectCardById(id);
+        if (c == null) {
+            return Results.getNotFoundResult("Card does not exist");
+        } else {
+            card.setId(id);
+            card.setUid(userService.getUserIdWithToken(token));
+            cardService.updateCard(card);
+            return Results.getSuccessResult();
+        }
+    }
+
+    @TokenRequired
+    @GetMapping("{id}")
+    public Result selectCardById(@RequestHeader("token") String token, @PathVariable("id") int id) {
+        Card c = cardService.selectCardById(id);
+        if (c == null) {
+            return Results.getNotFoundResult("Card does not exist");
+        } else {
+            if (c.getUid() == userService.getUserIdWithToken(token)) {
+                return Results.getSuccessResult(c);
+            }
+            else {
+                return Results.getNotFoundResult("Card does not exist");
+            }
+        }
     }
 }
